[FEATURE ember-application-engines] Refactor registry and resolver concerns for engines.
Provide a spare base implementation of `buildRegistry` for `Engine`,
which is extended by `Application`.

Also, allow for a custom `Resolver` property on `Engine` (and by
extension, `Application`).
